pyproject.toml
[tool.poetry]
name = "cdk-sample"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aws-cdk-lib = "^2.142.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
syrupy = "^4.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
